Better naming rules, merge webserver and capabilities into api
