+++ nyuki/api/api.py
@@ -1,8 +1,12 @@
 from aiohttp import web
+from aiohttp.hdrs import METH_ALL
+import asyncio
+from functools import partial
 import json
 import logging
 
 from nyuki.bus import reporting
+from nyuki.services import Service
 
 
 log = logging.getLogger(__name__)
@@ -40,7 +44,7 @@ class Response(web.Response):
         return kwargs.get('content_type') or kwargs.get('headers', {}).get('content_type')
 
 
-class WebServer(object):
+class WebServer:
 
     """
     The goal of this class is to gather all http-related behaviours.
@@ -142,3 +146,98 @@ async def mw_capability(app, capa_handler):
         return Response()
 
     return middleware
+
+
+def resource(path, versions=None, content_type=None):
+    """
+    Nyuki resource decorator to register a route.
+    A resource has multiple HTTP methods (get, post, etc).
+    """
+    def decorated(cls):
+        cls.RESOURCE_CLASS = ResourceClass(cls, path, versions, content_type)
+        return cls
+    return decorated
+
+
+class ResourceClass:
+
+    """
+    Allow the extensivity of the nyuki's HTTP resources using the webserver's
+    router from the `Api` class.
+    """
+
+    def __init__(self, cls, path, versions, content_type):
+        self.cls = cls
+        self.path = path
+        self.versions = versions
+        self.content_type = content_type
+
+    def _add_routes(self, router, path):
+        resource = router.add_resource(path)
+        cls_instance = self.cls()
+        for method in METH_ALL:
+            handler = getattr(self.cls, method.lower(), None)
+            if handler is not None:
+                handler = asyncio.coroutine(partial(handler, cls_instance))
+                handler.CONTENT_TYPE = self.content_type
+                route = resource.add_route(method, handler)
+                log.debug('Added route: %s', route)
+
+    def register(self, nyuki, router):
+        self.cls.nyuki = nyuki
+        if not self.versions:
+            self._add_routes(router, self.path)
+        else:
+            for version in self.versions:
+                route = '/{}{}'.format(version, self.path)
+                self._add_routes(router, route)
+
+
+class Api(Service):
+
+    """
+    Manage a webserver built using the nyuki's defined HTTP resources
+    """
+
+    CONF_SCHEMA = {
+        "type": "object",
+        "required": ["api"],
+        "properties": {
+            "api": {
+                "type": "object",
+                "properties": {
+                    "host": {"type": "string"},
+                    "port": {"type": "integer"}
+                }
+            }
+        }
+    }
+
+    def __init__(self, nyuki):
+        self._nyuki = nyuki
+        self._nyuki.register_schema(self.CONF_SCHEMA)
+        self._loop = self._nyuki.loop or asyncio.get_event_loop()
+        self._webserver = WebServer(self._loop)
+        self.host = None
+        self.port = None
+
+    @property
+    def capabilities(self):
+        return self._nyuki.HTTP_RESOURCES
+
+    async def start(self):
+        """
+        Expose capabilities by building the HTTP server.
+        The server will be started with the event loop.
+        """
+        for resource in self._nyuki.HTTP_RESOURCES:
+            resource.RESOURCE_CLASS.register(self._nyuki, self._webserver.router)
+        await self._webserver.build(self.host, self.port)
+
+    def configure(self, host='0.0.0.0', port=5558, debug=False):
+        self.host = host
+        self.port = port
+        self._webserver.debug = debug
+
+    async def stop(self):
+        await self._webserver.destroy()
